move getFilename to script
The logic of filename should be belong to script
not on the Snap. Right now Snap manages two object
1. Script
2. FakeServer
Now FakeServer needs filename, because it's directly
write to the file, and not through the script. If
there are a times, then we need to write the script
through the script. This will make, read and write format
should be in script. And if we think this is to big,
then we could separate again between script and reader/writer