fix panic when proxy log after test finish
As a proxy, there are function that log, when something is happened
Now when we are using dockertest when run the test, sometimes
the postgres container is finish after the test is done.

And this is call will cause panic cause it will call log in *testing.T
This commit will check isDone() first before logging. Cause disconnect
after test is finish is kind expected.